pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "macrokit-datalake"
version = "0.1.0"
description = "MacroKit data lake and ingestion tools"
requires-python = ">=3.8"
dependencies = [
    # Core DBT
    "dbt-core==1.8.0",
    "dbt-duckdb==1.8.0",
    # Database
    "duckdb==0.10.0",
    # Data fetching and processing
    "fredapi==0.5.1",
    "pandas==2.2.0",
    "numpy==1.24.3",
    # Configuration and environment
    "python-dotenv==1.0.0",
    "pyyaml==6.0.1",
    # CLI and logging
    "absl-py==2.1.0",
    # Essential utilities
    "requests==2.31.0",
    # Viz
    "streamlit",
    "click==8.0.0",
    "watchdog",
    "streamlit-plotly-events"
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    # Code quality
    "black==23.11.0",
    "flake8==6.1.0",
    "isort==5.12.0",
    "mypy==1.7.0",
    # Documentation
    "sphinx==7.2.6",
    "sphinx-rtd-theme==1.3.0",
    # Development tools
    "pre-commit==3.5.0",
    "jupyter==1.0.0",
    "ipython==8.17.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["macrokit_datalake*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false